eth: bring up on stm32, dhcp has some problem

+++ com/as.application/board.stm32f107vc/common/Can_Cfg.h
@@ -74,7 +74,7 @@ typedef enum {
 
 typedef struct {
 	void (*CancelTxConfirmation)( const Can_PduType *);
-	void (*RxIndication)( uint8 ,Can_IdType ,uint8 , const uint8 * );
+	void (*RxIndication)( uint16 ,Can_IdType ,uint8 , const uint8 * );
 	void (*ControllerBusOff)(uint8);
 	void (*TxConfirmation)(PduIdType);
 	void (*ControllerWakeup)(uint8);
